Update native harmony Text component to use harmony theme values
