+++ packages/harmony/src/components/text/Text/index.ts
@@ -1,2 +1,3 @@
 export { Text } from './Text'
 export * from './types'
+export * from './constants'
