+++ packages/mobile/src/app/index.ts
@@ -1,3 +1,3 @@
 export * from './App'
-// // Uncomment to run app in storybook mode
+// Uncomment to run app in storybook mode
 // export { default as App } from '../../.storybook'
